Don't throw cachefile away on the slightest error.
The cachefile used to be ignored, deleted and expensively rebuild if even one file was off. Since macOS for example throws .DS_Store files everywhere this used to happend a lot. Now the cachefile will be read and compared with the actual folder and new files will be added to the index and files that cannot be found anymore removed from it.