Removed faulty function mkdir
The function tried finding existing folders but failed to do so properly.
First it had issues with umlauts because the path name was not beeing normalized.
Then to allow for foldername changes it tried to match the foldername with startsWith.
This worked but alas reported false positives, like "10" for "1".

Thus now a hidden file gets created in every folder for matching them, which allows for complete control over the foldername.